staff/requirements.txt
python-telegram-bot==20.7
groq
gspread
oauth2client
python-dotenv